CORE-7115 Move UtxoLedgerTransactionImpl to libs/ledger/ledger-utxo-data
