+++ components/ledger/ledger-utxo-flow/build.gradle
@@ -13,6 +13,7 @@ dependencies {
     implementation "net.corda:corda-cipher-suite"
     implementation project(":libs:crypto:crypto-core")
     implementation project(":libs:ledger:ledger-common-data")
+    implementation project(":libs:ledger:ledger-utxo-data")
     implementation project(":libs:sandbox")
 
     testImplementation "org.mockito.kotlin:mockito-kotlin:$mockitoKotlinVersion"
