+++ components/ledger/ledger-utxo-flow/src/main/kotlin/net/corda/ledger/utxo/impl/transaction/UtxoSignedTransactionImpl.kt
@@ -1,6 +1,7 @@
 package net.corda.ledger.utxo.impl.transaction
 
 import net.corda.ledger.common.data.transaction.WireTransaction
+import net.corda.ledger.utxo.data.transaction.UtxoLedgerTransactionImpl
 import net.corda.v5.application.crypto.DigitalSignatureAndMetadata
 import net.corda.v5.application.crypto.DigitalSignatureVerificationService
 import net.corda.v5.application.crypto.SigningService
