+++ components/ledger/ledger-utxo-flow/src/main/kotlin/net/corda/ledger/utxo/impl/transaction/UtxoTransactionBuilderImpl.kt
@@ -5,7 +5,9 @@ import net.corda.ledger.common.data.transaction.PrivacySaltImpl
 import net.corda.ledger.common.data.transaction.TransactionMetaData
 import net.corda.ledger.common.data.transaction.WireTransaction
 import net.corda.ledger.common.data.transaction.createTransactionSignature
-import net.corda.ledger.utxo.impl.state.TransactionStateImpl
+import net.corda.ledger.utxo.data.transaction.UtxoComponentGroup
+import net.corda.ledger.utxo.data.transaction.UtxoOutputInfoComponent
+import net.corda.ledger.utxo.data.state.TransactionStateImpl
 import net.corda.ledger.utxo.impl.timewindow.TimeWindowBetweenImpl
 import net.corda.ledger.utxo.impl.timewindow.TimeWindowUntilImpl
 import net.corda.sandbox.SandboxGroup
