+++ libs/ledger/ledger-utxo-data/build.gradle
@@ -1,0 +1,16 @@
+plugins {
+    id 'corda.common-publishing'
+    id 'corda.common-library'
+}
+
+description 'Corda ledger UTXO data'
+
+dependencies {
+    compileOnly 'org.osgi:osgi.annotation'
+    compileOnly 'org.osgi:org.osgi.service.component.annotations'
+
+    implementation "net.corda:corda-ledger-utxo"
+
+    implementation project(":libs:crypto:crypto-core")
+    implementation project(":libs:ledger:ledger-common-data")
+}
