+++ libs/ledger/ledger-utxo-data/src/main/kotlin/net/corda/ledger/utxo/data/state/TransactionStateImpl.kt
@@ -1,4 +1,4 @@
-package net.corda.ledger.utxo.impl.state
+package net.corda.ledger.utxo.data.state
 
 import net.corda.v5.ledger.common.Party
 import net.corda.v5.ledger.utxo.Contract
