+++ settings.gradle
@@ -239,6 +239,7 @@ include 'libs:layered-property-map:layered-property-map-test-converter'
 include 'libs:ledger:notary'
 include 'libs:ledger:ledger-common-data'
 include 'libs:ledger:ledger-consensual-data'
+include 'libs:ledger:ledger-utxo-data'
 include 'libs:lifecycle:lifecycle'
 include 'libs:lifecycle:lifecycle-impl'
 include 'libs:lifecycle:lifecycle-test-impl'
