+++ testing/ledger/ledger-utxo-testkit/src/main/kotlin/net/corda/ledger/utxo/testkit/UtxoInvalidState.kt
@@ -1,7 +1,7 @@
 package net.corda.ledger.utxo.testkit
 
-import net.corda.ledger.utxo.impl.state.StateAndRefImpl
-import net.corda.ledger.utxo.impl.state.TransactionStateImpl
+import net.corda.ledger.utxo.data.state.StateAndRefImpl
+import net.corda.ledger.utxo.data.state.TransactionStateImpl
 import net.corda.v5.crypto.SecureHash
 import net.corda.v5.ledger.utxo.StateRef
